+++ modules/behavior.js
@@ -82,6 +82,11 @@ _extend(Behavior.prototype, CommonMixin, DelegateEntityEventsMixin, UIMixin, Vie
     this._undelegateViewEvents();
 
     this.el = this.view.el;
+    if (this.view.$el) {
+      this.$el = this.view.$el;
+    } else {
+      delete this.$el;
+    }
 
     this._delegateViewEvents(this.view);