+++ modules/view.js
@@ -47,6 +47,7 @@ function childReducer(children, region) {
 const View = function(options) {
   this.cid = uniqueId(this.cidPrefix);
   this._setOptions(options, ClassOptions);
+  this.Dom = _extend({}, this.Dom);
 
   this.preinitialize.apply(this, arguments);
 
@@ -73,7 +74,15 @@ _extend(View.prototype, ViewMixin, RegionsMixin, {
 
   setElement(element) {
     this._undelegateViewEvents();
-    this.el = this._validateEl(element);
+    const el = this._validateEl(element);
+    const wrappedEl = this.Dom.wrapEl && this.Dom.wrapEl(el);
+
+    this.el = el;
+    if (this.Dom.wrapEl) {
+      this.$el = wrappedEl;
+    } else {
+      delete this.$el;
+    }
     this._setBehaviorElements();
 
     this._isRendered = this.Dom.hasContents(this.el);